Restructured project so ui components are in dedicated subpackage.
