+++ lib/ui/sub_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:leaf_for_reddit/official_themes.dart';
 import 'package:transparent_image/transparent_image.dart';
 
-import 'official_themes.dart';
 
 class SubList extends StatefulWidget {
   @override
